Refactor: Organize marketing components into dedicated folder structure
• Created new components/marketing/ directory with organized subdirectories
• Moved marketing-specific components from ui/ to marketing/ with proper categorization:
  - sections/ for page sections (benefits, testimonials, CTAs, etc.)
  - effects/ for visual effects (lamp, marquee, staggered-text)
  - forms/ for marketing forms (waitlist-form)
  - buttons/ for marketing buttons (hover-button)
  - navigation/ for navbar component
• Updated all import statements across 7 pages and 2 components
• Fixed internal component imports after reorganization
• Maintained separation between core UI components and marketing components
• All tests passing, TypeScript compilation successful

This improves codebase organization by clearly separating marketing-specific
components from the base UI component library.
