+++ app/login/page.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 import Link from 'next/link';
 import { usePostHog } from "posthog-js/react";
-import { HoverButton } from "@/components/ui/hover-button";
+import { HoverButton } from "@/components/marketing/buttons/hover-button";
 import { Input } from "@/components/ui/input";
 import { Form, FormControl, FormField, FormItem, FormMessage } from "@/components/ui/form";
 import { motion, AnimatePresence } from "framer-motion";
