+++ components/marketing/effects/lamp-effect.tsx
