+++ components/marketing/effects/marquee.tsx
