+++ components/marketing/effects/text-animate.tsx
